#  mat: rewrite Cholesky.ToSym without calling SymOuterK